+++ src/template/README.md
@@ -7,5 +7,5 @@ sdk: gradio
 sdk_version: 4.5.0
 app_file: app.py
 pinned: True
-license: MIT
+license: mit
 ---
